% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic.R
\name{aic}
\alias{aic}
\title{Akaike Information Criterion Score Calculation}
\usage{
aic(preds, obs, numVar)
}
\arguments{
\item{preds}{Predicted Values}

\item{obs}{Observed or Actual Values}
}
\value{
AIC Score
}
\description{
Akaike Information Criterion Score Calculation
}
\examples{
aic(c(1,1,1),c(1,0.5,1))
}
