% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Arear Under Curve Score Calculation}
\usage{
auc(preds, obs)
}
\arguments{
\item{preds}{Predicted Values}

\item{obs}{Observed or Actual Values}
}
\value{
AUC Score
}
\description{
Arear Under Curve Score Calculation
}
\examples{
auc(c(1,1,1),c(1,0.5,1))
}
