% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{bic}
\alias{bic}
\title{Bayesian Information Criterion Score Calculation}
\usage{
bic(preds, obs, numVar)
}
\arguments{
\item{preds}{Predicted Values}

\item{obs}{Observed or Actual Values}
}
\value{
BIC Score
}
\description{
Bayesian Information Criterion Score Calculation
}
\examples{
bic(c(1,1,1),c(1,0.5,1))
}
