% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binner.R
\name{binner}
\alias{binner}
\title{Contineous Variable Binner}
\usage{
binner(x, y = NULL, lower, upper, by, min.obs, sep = "-",
  above.char = "+", digits)
}
\arguments{
\item{x}{Variable to bin}

\item{y}{Conditional binnig on depedent variable}

\item{lower}{Lower bound of bins}

\item{upper}{Uper bound of bins}

\item{by}{speration between bins}

\item{min.obs}{Minimum Observation required in each bin}

\item{sep}{Bin bound seperator}

\item{above.char}{Charcter to denote value beyond upper bound}

\item{digits}{No of digits to be considered for binning}
}
\value{
A binned categorical vector of x
}
\description{
Contineous Variable Binner
}
\examples{
binner(c(1,2,3,4,5,6,7,8,9,10,11,13,15,77,99))
vec <- c(1,1,1,1,1,2,2,2,2,2,2,5,5,5,6,6,7,8,9,9,10,11,13,15,77,99)
binner(vec)
binner(vec,by=7)
binner(vec,by=3)
binner(vec,by=3,upper=20)

y <- sample(0:1,NROW(vec_),replace = T)
binner(vec,y)
binner(seq(0,100,1),by=12)
}
