% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyColGen.R
\name{dummyColGen}
\alias{dummyColGen}
\title{The Dummy Column Generator}
\usage{
dummyColGen(data, name, keep.column = TRUE, keep.col.name = TRUE,
  fill.value = FALSE, fill.value.from = NULL)
}
\arguments{
\item{data}{data containing the column for which dummy columns to be generated}

\item{name}{Name of Column}

\item{keep.column}{To keep column within the data}

\item{keep.col.name}{keep column name in dummy colmn}

\item{fill.value}{Wheteher fill value from some other column}

\item{fill.value.from}{Column name from whihc value to be stored in the dummy columns}
}
\value{
a data frame with dummy columns
}
\description{
The Dummy Column Generator
}
