% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyGain.R
\name{entropyGain}
\alias{entropyGain}
\title{Entropy Gain of a Categorical Variable}
\usage{
entropyGain(target, variable)
}
\arguments{
\item{target}{Depedent Variable}

\item{variable}{Indepedent Variable}
}
\value{
Entropy Gain Score
}
\description{
Entropy Gain of a Categorical Variable
}
\examples{
variable=c("A","A","B","A","B","B","B","C","C","C","C","B",
 "B","B","A","A","C","C","C","C")
target = c(1,0,0,0,2,1,1,2,1,2,2,1,1,0,1,2,1,0,2,1)
entropyGain(target,variable)

}
