% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoCategory.R
\name{infoCategory}
\alias{infoCategory}
\title{Information for Categorical Variables}
\usage{
infoCategory(data, numFactors = 20)
}
\arguments{
\item{data}{A dat frame with variables}

\item{numFactors}{possible number of factors for a character variables}
}
\value{
a data frame with summary for categorical variables
}
\description{
Information for Categorical Variables
}
\examples{

df <- data.frame(NormalFactor = as.factor(sample(LETTERS[2:10],100,replace = T)),
 VeryLargeFactors = as.factor(sample(LETTERS,100,replace = T)),
 NormalCharacters = sample(LETTERS[2:20],100,replace = T),
 VeryLargeCharacters = sample(c(LETTERS,letters),100,replace = T))

infoCategory(df)
infoCategory(df,length(c(LETTERS,letters)))

}
