% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoContinuous.R
\name{infoContinuous}
\alias{infoContinuous}
\title{Continuous variable information}
\usage{
infoContinuous(data, var.specifier = "", transpose = FALSE)
}
\arguments{
\item{data}{A data frame for which information is to be retireved}

\item{var.specifier}{Additional variable name specifier}

\item{transpose}{Get a transpose of a resultant information.
Not Suitable for very large number of variables}
}
\value{
a dta frame with information
}
\description{
Continuous variable information
}
\examples{
data(mtcars)
infoContinuous(mtcars)

}
