% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informationValue.R
\name{informationValue}
\alias{informationValue}
\title{Information Value of a Contineous Variable}
\usage{
informationValue(target, variable, event, upper, by)
}
\arguments{
\item{target}{Depedent Variable}

\item{variable}{Indepedent Variable}

\item{upper}{Upper bound of the bins}

\item{by}{Binning difference of contieous variable}
}
\value{
Information Value of the Variable
}
\description{
Information Value of a Contineous Variable
}
\examples{
vec <- c(1,1,1,1,1,2,2,2,2,2,2,5,5,5,6,6,7,8,9,9,10,11,13,15,77,99)
y <- sample(0:1,NROW(vec_),replace = T)

informationValue(target = y,variable = vec)
informationValue(target = y,variable = vec,by=2)
informationValue(target = y,variable = vec,by=1)
informationValue(target = y,variable = vec,by=4)
informationValue(target = y,variable = vec,by=3)
informationValue(target = y,variable = vec,by=3,upper = 50)
informationValue(target = y,variable = vec,by=3,upper = 20)
informationValue(target = y,variable = vec,by=3,upper = 15)
informationValue(target = y,variable = vec,by=3,upper = 12)
informationValue(target = y,variable = vec,by=3,upper = 9)
informationValue(target = y,variable = vec,by=3,upper = 6)
}
