% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Wrapper for loading multiple files in coherent manner within a folder.}
\usage{
loadData(path, patterns, lib.selector = "readr", verbose = FALSE,
  return.df = T, start.at, recursive = FALSE)
}
\arguments{
\item{path}{Path Of the folder in which files are stored}

\item{patterns}{Pattern of filename to be loaded}

\item{lib.selector}{"readr" or "data.table" to be used}

\item{verbose}{print coherent output}

\item{return.df}{return a data.frame or list of many data.frames}

\item{recursive}{Recursive search for pattern in th folder}
}
\value{
a Data Frame or a List of Data Frames.
}
\description{
Wrapper for loading multiple files in coherent manner within a folder.
}
