% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error Score Calculation}
\usage{
mae(preds, obs)
}
\arguments{
\item{preds}{Predicted Values}

\item{obs}{Observed or Actual Values}
}
\value{
MAE Score
}
\description{
Mean Absolute Error Score Calculation
}
\examples{
mae(c(1,1,1),c(1,0.5,1))
}
