% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeRankedMatrix.R
\name{MakeRankedMatrix}
\alias{MakeRankedMatrix}
\title{Making varable selection based ranking matirx}
\usage{
MakeRankedMatrix(variableMatrix, candidateColumn = NULL)
}
\arguments{
\item{variableMatrix}{A data frame with variable name and scored methods}

\item{candidateColumn}{The column which should be treated as column
for candidates. If NULL first clumn will be treated as candidate column.}
}
\value{
A Ranked Matrix
}
\description{
Making varable selection based ranking matirx
}
