% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape.R
\name{mape}
\alias{mape}
\title{Mean Absolute Percentage Error Score Calculation}
\usage{
mape(preds, obs)
}
\arguments{
\item{preds}{Predicted Values}

\item{obs}{Observed or Actual Values}
}
\value{
MAPE Score
}
\description{
Mean Absolute Percentage Error Score Calculation
}
\examples{
mape(c(1,1,1),c(1,0.5,1))
}
