% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Normalise the Vairables}
\usage{
normalise(x, y = NULL, method = "zscore", na.rm = TRUE, diff.val = 0,
  norm.list = list(), depedentMod = NULL, scores = NULL)
}
\arguments{
\item{x}{A vector or Data frame to be normalised}

\item{y}{A depedent vector for depedent normalisation.
Default value is NULL(indepedent normalisation).}

\item{method}{A method or a vector of methods to apply normalisation on varaibles.
If vector not equal to num of variables "unit" normalisation is implemented for
the rest. Methods implemented are
1) unitMean
2) unitSD
3) unitMeanSD
4) minMax
5) boxCox
6) log
7) meanDiff
8) medianDiff
9) meanSDDiff
10) medianSDDiff
11) zscore (zero Mean and unit SD)
12) power}

\item{na.rm}{A boolean for na treatment. Default is TRUE.}

\item{diff.val}{Difference required from mean and median}

\item{norm.list}{a list of vlaues for flexible normalisation}

\item{depedentMod}{Model to be used for Dependent Normalisation(lm or glm)}

\item{scores}{scores to be used for evaluting the normalisation}
}
\value{
Normalised vector or a data frame
}
\description{
Normalise the Vairables
}
\details{
log is impleted with 1 added to the value
}
\examples{
normalise(mtcars)

}
