% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predScores.R
\name{predScores}
\alias{predScores}
\title{Prediction Scoring}
\usage{
predScores(preds, obs, numVar = NULL)
}
\arguments{
\item{preds}{Predicted Values}

\item{obs}{Observed or Actual Values}

\item{numVar}{Number of predictors used to build the model.
If NULL, general R squared is calculated with no AIC and BIC.}
}
\value{
Prediction scores
}
\description{
Prediction Scoring
}
\examples{
predScores(c(1,1,1),c(1,0.5,1))

}
