% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error Score Calculation}
\usage{
rmse(preds, obs)
}
\arguments{
\item{preds}{Predicted Values}

\item{obs}{Observed or Actual Values}
}
\value{
RMSE Score
}
\description{
Root Mean Square Error Score Calculation
}
\examples{
rmse(c(1,1,1),c(1,0.5,1))
}
