% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquared.R
\name{rsquared}
\alias{rsquared}
\title{R squared and Adjusted R squared Score Calculation}
\usage{
rsquared(preds, obs, numVar = NULL)
}
\arguments{
\item{preds}{Predicted Values}

\item{obs}{Observed or Actual Values}

\item{numVar}{Number of predictors used to build the model.
If NULL general R squared is calculated.}
}
\value{
R-squared / Adjusted R-Squared Score
}
\description{
R squared and Adjusted R squared Score Calculation
}
\examples{
rsquared(c(1,1,1),c(1,0.5,1))
rsquared(c(1,1,1),c(1,0.5,1),numVar = 1)
}
